# api/openapi.yaml
openapi: 3.0.3
info:
  title: FlowHook API
  description: |
    FlowHook is a webhook testing and inspection platform. This API allows you to:
    - Create and manage webhook endpoints
    - Capture and inspect incoming webhooks
    - Replay requests to other URLs
    - Configure auto-forwarding rules
    - Transform requests/responses
    - View analytics and delivery statistics
  version: 1.0.0
  contact:
    name: FlowHook Support
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://api.flowhook.com
    description: Production server
...
paths:
...
  /api/v1/endpoints:
    post:
      summary: Create endpoint
      description: Creates a new webhook endpoint
      tags:
        - Endpoints
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: My Webhook Endpoint
      responses:
        '201':
          description: Endpoint created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Endpoint'
    get:
      summary: List endpoints
      description: Returns a list of all endpoints
      tags:
        - Endpoints
      responses:
        '200':
          description: List of endpoints
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Endpoint'
...
  /api/v1/auth/register:
    post:
      summary: Register user
      description: Creates a new user account
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
                name:
                  type: string
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
  /api/v1/auth/login:
    post:
      summary: Login
      description: Authenticates a user and returns a session token
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
components:
  schemas:
    Endpoint:
      type: object
      properties:
        id:
          type: string
          format: uuid
        slug:
          type: string
        name:
          type: string
        created_at:
          type: string
          format: date-time
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        name:
          type: string
        created_at:
          type: string
          format: date-time
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - BearerAuth: []